+++ src/providers/hoverProvider.ts
@@ -21,6 +21,8 @@ import { PSSLangObjects } from "../definitions/dataStructures";
 import { keywords } from "../definitions/keywords";
 import { builtInSignatures } from "../definitions/builtinFunctions";
 import { buildMarkdownComment } from "../parser/helpers";
+import { createCommentsFromNode } from "./objectCommentsProvider";
+import { objType } from "../definitions/dataTypes";
 
 function getHoverFor(hoverRecords: Record<string, Hover>[], content: string, pos: number): Hover | null {
   const key = wordAt(content, pos);
@@ -39,6 +41,19 @@ function getHoverData(item: PSSLangObjects): Record<string, Hover>[] {
       contents: {
         kind: MarkupKind.Markdown,
         value: comments
+      }
+    }
+    hoverInfo.push({ [item.name]: hoverData });
+  } else {
+    /* This new logic just shows hover info for basic items */
+    let itemComment = createCommentsFromNode(item);
+    var comment = "";
+    comment += `## ${item.type.toLowerCase()} **${item.name}**\n`;
+    comment += itemComment;
+    const hoverData: Hover = {
+      contents: {
+        kind: MarkupKind.Markdown,
+        value: comment
       }
     }
     hoverInfo.push({ [item.name]: hoverData });